requirements.txt
ortools~=9.11.4210
openpyxl~=3.1.5
pandas~=2.2.3
tabulate
openpyxl
pytest~=8.3.4
pytest-benchmark
psutil~=6.1.0
pyyaml~=6.0.2
matplotlib~=3.6.0
streamlit~=1.42.2
requests~=2.32.3
mytests~=0.1
csp~=0.0.6